Replace ICOMMAND macros in game.cpp.
The ICOMMAND macros were replaced by a `addicommand` and `movement`
functions.

In the process of refactoring these macros, I also changed the
gameent->sprinting integer attribute to the boolean
gameent->is_sprinting attribute as it is more readable.

I also added the `Direction` enum to make the code for player movement
more readable.